Update vendored zstd from 1.3.8 to 1.4.0.

Quick bench (macOS):
```
$ benchstat old.txt new.txt
name                    old time/op   new time/op   delta
StreamCompression-12     91.8ms ± 2%   91.8ms ± 1%    ~     (p=1.000 n=5+4)
StreamDecompression-12   15.9ms ± 2%   16.5ms ± 2%  +3.40%  (p=0.008 n=5+5)
Compression-12           90.8ms ± 1%   92.2ms ± 3%    ~     (p=0.222 n=5+5)
Decompression-12         13.5ms ± 2%   13.8ms ± 1%    ~     (p=0.056 n=5+5)

name                    old speed     new speed     delta
StreamCompression-12    109MB/s ± 2%  109MB/s ± 1%    ~     (p=1.000 n=5+4)
StreamDecompression-12  626MB/s ± 2%  605MB/s ± 2%  -3.29%  (p=0.008 n=5+5)
Compression-12          110MB/s ± 1%  108MB/s ± 3%    ~     (p=0.222 n=5+5)
Decompression-12        741MB/s ± 2%  725MB/s ± 1%    ~     (p=0.056 n=5+5)
```